Zmniejszenie wachań dzięki exponential moving average
For averaging, you can keep a running average of the last n frequencies by maintaining a list of the last n frequencies and updating the average each time a new frequency is measured.

For exponential moving average, you can use the current frequency and the previous exponential moving average to calculate the new exponential moving average. This can be done using the following formula:

EMA = alpha * current_frequency + (1 - alpha) * previous_EMA

where alpha is the smoothing factor, which determines the weight given to the current frequency. A value of alpha closer to 1 gives more weight to the current frequency, while a value closer to 0 gives more weight to the previous EMA.